refactor(mountains): extract noise primitives into mountains/noise.ts
Move SimplexNoise, the deterministic fixed-seed fBm helpers (hashLattice /
valueNoise / fbm), the aperiodic terrainRidgeField, and forestDensityField into
src/mountains/noise.ts (pure math, no THREE/DOM). The hub imports them and
re-exports terrainRidgeField/forestDensityField to preserve the public
./mountains.js named exports and the terrain-tests.js direct import.

Byte-identical: terrain (9/9), regression (10/10), verify (49/49), tsc clean.
